(subtask) - Add delete all subtasks and confirm

+++ task_manager.py
@@ -247,7 +247,13 @@ class TaskManager:
                 print(messages.ALL_TASKS_DELETED_MESSAGE)
 
         elif isinstance(task, Task):
-            self.delete_all_subtasks(task)
+            if self.has_subtasks(task):
+                print(messages.CONFIRM_DELETE_IF_HAS_SUBTASKS_MESSAGE)
+                confirm = input()
+                if confirm.lower() == "y":
+                    self.delete_all_subtasks(task)
+                else:
+                    return
             self.session.delete(task)
             self.session.commit()
             print(messages.TASK_DELETED_MESSAGE.format(deleted_task=task.title))
@@ -256,13 +262,20 @@ class TaskManager:
             raise ValueError(messages.INVALID_CHOICE_MESSAGE)
     
     #TODO: if 'ALL' entered as an input then delete all the subtasks of the task 
-    def delete_subtask(self, get_task):
+    def delete_subtask(self, get_task=None):
         print(messages.ENTER_DELETE_SUBTASK_MESSAGE)
-        subtask = self.get_task(get_task=get_task)
-        self.delete_all_subtasks(subtask)
-        self.session.delete(subtask)
-        self.session.commit()
-        print(messages.TASK_DELETED_MESSAGE.format(deleted_task=subtask.title))
+        subtask = self.get_task(commands=True, get_task=get_task)
+
+        if subtask == app_commands.get('ALL'):
+            self.delete_all_subtasks(get_task)
+            print(messages.ALL_SUBTASKS_DELETED_MESSAGE)
+
+        else:
+            self.delete_all_subtasks(subtask)
+            self.session.delete(subtask)
+            self.session.commit()
+            print(messages.TASK_DELETED_MESSAGE.format(deleted_task=subtask.title))
+        
 
     def delete_all_subtasks(self, task):
         subtasks = self.session.query(Task).filter(Task.parent_id == task.id).all()
